Exclude a repo from configuration management
This adds the ability to exclude a (core) repo from configuration
management, and uses that ability to exclude opensafely-core/ethelred.

Why?

Team PIN have been experimenting with trunk-based development as well as
GitHub flow workflows, and have decided to adopt a workflow that sits
between the two: [Ship / Show / Ask][1]. To adopt this workflow, they
would like to allow PRs to be merged to the mainline branch without
approval; and *possibly* to remove branch protection rules from the
mainline branch.

[1]: https://martinfowler.com/articles/ship-show-ask.html